requirements.txt
ultralytics
pycocotools
python-dotenv
comet-ml
numpy
pyyaml